WIP Add some trace to generic_cpp_std_unordered_map_to_py_dict as it was asserting with higher than expected reference counts for what looks like interned longs.
The asserts do not happen if large numbers are used such as 1000 onwards.
When using 0 onwards the reference counts after PyDict_SetItem can be two or three higher not just one higher.
This suggests that our reference counting tests can be wrong with interned numbers.
Is it possible a context switch is happening in PyDict_SetItem and another thread increases the refcount of an interned number?